Allow one object-server per disk deployment
Enabled by a config boolean, "server_per_disk" in the
"app:object-server" config section.  In this mode, the parent
swift-object-server process continues to run as root, and forks off one
worker per unique port defined in all the rings for the server.  The
parent process detects dead servers and restarts them (on the correct
port), will start missing servers when an updated ring file is found
with a device on the server with a new port, and will kill extraneous
servers when their port is found to no longer be in the ring.

Immediately stopping all swift-object-worker processes still works by
sending the parent a SIGTERM (it continues to be propagated to the
children).  Likewise, a SIGHUP to the parent process will be propagated
to the children who will finish their current request(s) and then exit.

The class method RingData.load_devices() will load only the device array
from a serialized, gzipped ring file, retaining a reference to only the
returned array of device dicts.  This is used to keep the parent and all
forked off workers from having full copies of all storage policy rings
in memory, when they don't need it.

A new helper method,
swift.common.storage_policy.all_bind_ports_for_node(), returns a set of
all device ports in all rings for a server (identified by its set of IP
addresses).  The function includes support for caching mtimes of the
ring files, so they are not opened more frequently than necessary (if
no rings have changed mtimes, then None is returned).  This is obviously
used by swift-object-server when server_per_disk=True and takes
advantage of RingData.load_devices().

[#94449854]