Allow SAIO to answer is_local_device() better
Common SAIO setups have multiple "servers" all on the same host with
different ports for the different "servers" (which happen to match the
IPs specified in the rings for the devices on each of those "servers").

This breaks down if you want to set up the SAIO to have one
object-server per disk.

However, you can configure the SAIO to have different localhost IP
addresses (e.g. 127.0.0.1, 127.0.0.2, etc.) in the ring and in the
servers' config files' bind_ip setting.

This new whataremyips() implementation combined with a little plumbing
allows is_local_device() to accurately answer, even on an SAIO.

In the default case (an unspecified bind_ip defaults to '0.0.0.0') as
well as an explict "bind to everything" like '0.0.0.0' or '::',
whataremyips() behaves as it always has, returning all IP addresses for
the server.

Also updated probe tests to handle each "server" in the SAIO having a
unique IP address.

Change-Id: I7dfef9cc4a3c48145011d9d13d904e6b623de020