Make the reconstructor handoffs_first work (and useful)
The handoffs_first mode in the replicator has the useful behavior of
processing all handoff parts across all disks until there aren't any
handoffs anymore on the node [1] and then it seemingly tries to drop
back into normal operation.  In practice I've only ever heard of
handoffs_first used while rebalancing and turned off as soon as the
rebalance finishes - it's not recommended to run with handoffs_first
mode turned on and it emits a warning while the option is enabled.

The handoffs_first mode on the reconstructor doesn't work - it was
prioritizing handoffs *per-part* [2] - which is really unfortunate
because in the reconstructor during a rebalance it's often *much* more
attractive from an efficiency disk/network perspective to revert a
partition from a handoff than it is to rebuild an entire partition from
another primary using the other EC fragments in the cluster.

This change fixes handoffs_first so that it's useful - and just like
handoffs_first mode in the replicator - it gives the operator the option
of forcing the consistency engine focus on rebalance.

1. Ia324728d42c606e2f9e7d29b4ab5fcbff6e47aea fixed replicator
handoffs_first "mode"

2. Unlike replication each partition in a EC policy can have a different
kind of job per frag_index, but the cardinality of jobs is typically 1
for revert and 2 for sync.

Known-Issues:

Despite the config option name handoffs_first this change really
implements more of a handoffs_*only* mode.  This behavior is somewhat
consistent with the replicator's option (any error on any handoff will
make it essentially handoff only forever) but very true in the
reconstructor.  This implementation will start to log a warning every
cycle if you leave your reconstructor running in handoffs_first.

Closes-Bug: #1653018

Change-Id: Idde4b6cf92fab6c45f2c0c2733277701eb436898